Moved vacations synchronization to separate actor.
